refactoring to extract a callable function
