+++ src/benchmark/benchmark_bvh.cpp
@@ -1,0 +1,8 @@
+#include <benchmark/benchmark.h>
+
+static void BM_BVH_Tracing(benchmark::State& state) {
+
+}
+ 
+
+BENCHMARK_MAIN();
