+++ src/bvh/CMakeLists.txt
@@ -1,14 +1,16 @@
-add_library(bvhlib OBJECT  "boundable.cpp" "bvh.cpp")
-target_include_directories(bvhlib PUBLIC "../common/")
 find_package(OpenMP REQUIRED)
 
-add_executable(sphere_bvh_single_threaded "sphere_bvh.cpp")
 add_compile_options(-O3 -g)
-target_include_directories(sphere_bvh_single_threaded PUBLIC "../common/" "../data_porting" "./")
-# target_link_libraries(sphere_bvh_single_threaded bvhlib tracer_common shapeio nlohmann_json::nlohmann_json OpenMP::OpenMP_CXX)
-target_link_libraries(sphere_bvh_single_threaded bvhlib tracer_common nlohmann_json::nlohmann_json OpenMP::OpenMP_CXX)
+
 
 if(OpenMP_CXX_FOUND)
+
+# the bvh ray tracing library, including the ray tracing methods.
+  add_library(bvhlib OBJECT "ray_tracing.cpp" "boundable.cpp" "bvh.cpp")
+  target_include_directories(bvhlib PUBLIC "../common/")
+  target_link_libraries(bvhlib tracer_common OpenMP::OpenMP_CXX )
+
+# an executable for multiple thread bvh
   add_executable(bvh_mt "sphere_bvh_openmp.cpp")
   target_include_directories(bvh_mt PUBLIC "../common/" "../data_porting" "./")
   target_link_libraries(bvh_mt bvhlib tracer_common shapeio nlohmann_json::nlohmann_json OpenMP::OpenMP_CXX)
@@ -24,10 +26,10 @@ target_link_libraries(
   bvh_test
   gtest_main
   bvhlib
+  tracer_common
 )
 target_include_directories(
   bvh_test PUBLIC "../common")
-
 include(GoogleTest)
 gtest_discover_tests(bvh_test)
